Rename certau.lib to certau.util
