+++ certau/util/stix/helpers.py
