+++ certau/util/taxii/util.py
